src/requirements.txt
Django
coverage
tox
gunicorn
psycopg2
django-dotenv
teamcity-messages
